Still some serious bugs, but I'm fixing them.
There's a lot of pitfalls in Java 8->7 backporting; notably, the same
lambda is equivalent to that lambda if created in the same way, but two
anonymous inner classes created independently will have different
references (and they can't implement equals() and still be functional
interfaces for people who do use Java 8!). This was tricky but seems to
be resolved in some of the tests. Others are going to be harder; the
regex tests had some inconsistencies that when corrected made the tests
succeed; I'm not sure if "change zero-length strings to the character
'x'" was the intent of the original tests.