Marker: fails less tests than original
I spent a lot of time trying to resolve a peculiar bug in code called by
IntTest, only to find the original codebase failed at the same place --
as well as failing in BuilderCacheTest, which mine doesn't! So this
should be a good starting point. All tests pass currently, though it's
not clear if the behavior in CompactIntSubset is correct (sizes don't
always match, in this or in the original code, which makes an assertion
fail). Next up is improving Unicode support, using code I wrote for
RegExodus.